+++ core/network/helpers/protobuf_message_read_writer.hpp
@@ -14,23 +14,28 @@
 #include <outcome/outcome.hpp>
 
 #include "network/helpers/message_read_writer.hpp"
-#include "network/helpers/adapters.hpp"
+#include "network/adapters/uvar.hpp"
+#include "network/adapters/protobuf.hpp"
 #include "scale/scale.hpp"
 
 namespace kagome::network {
   /**
    * Read and write messages, encoded into protobuf with a prepended varint
    */
-  class ProtobufMessageReadWriter : public std::enable_shared_from_this<ProtobufMessageReadWriter> {
+  class ProtobufMessageReadWriter
+      : public std::enable_shared_from_this<ProtobufMessageReadWriter> {
     template <typename MsgType>
     using ReadCallback = std::function<void(outcome::result<MsgType>)>;
 
    public:
-    explicit ProtobufMessageReadWriter(std::shared_ptr<libp2p::basic::MessageReadWriter> read_writer)
+    explicit ProtobufMessageReadWriter(
+        std::shared_ptr<libp2p::basic::MessageReadWriter> read_writer)
         : read_writer_{std::move(read_writer)} {}
-    explicit ProtobufMessageReadWriter(const std::shared_ptr<libp2p::basic::ReadWriter> &read_writer)
-        : read_writer_{std::make_shared<libp2p::basic::MessageReadWriterUvarint>(
-        read_writer)} {}
+    explicit ProtobufMessageReadWriter(
+        const std::shared_ptr<libp2p::basic::ReadWriter> &read_writer)
+        : read_writer_{
+            std::make_shared<libp2p::basic::MessageReadWriterUvarint>(
+                read_writer)} {}
 
     /**
      * Read a Protobuf message from the channel
@@ -39,21 +44,25 @@ namespace kagome::network {
      */
     template <typename MsgType>
     void read(ReadCallback<MsgType> cb) const {
-      read_writer_->read(
-          [self{shared_from_this()}, cb = std::move(cb)](auto &&read_res) {
-            if (!read_res) {
-              return cb(read_res.error());
-            }
+      read_writer_->read([self{shared_from_this()},
+                          cb = std::move(cb)](auto &&read_res) {
+        if (!read_res) {
+          return cb(read_res.error());
+        }
 
-            using ProtobufRW = MessageReadWriter<ProtobufMessageAdapter<MsgType>, NoSink>;
-            using UVarRW = MessageReadWriter<UVarMessageAdapter<MsgType>, ProtobufRW>;
+        using ProtobufRW =
+            MessageReadWriter<ProtobufMessageAdapter<MsgType>, NoSink>;
+        using UVarRW =
+            MessageReadWriter<UVarMessageAdapter<MsgType>, ProtobufRW>;
 
-            MsgType msg;
-            if (auto msg_res = UVarRW::read(msg, *read_res.value(), read_res.value()->begin()); !msg_res) {
-              return cb(msg_res.error());
-            }
-            return cb(std::move(msg));
-          });
+        MsgType msg;
+        if (auto msg_res =
+                UVarRW::read(msg, *read_res.value(), read_res.value()->begin());
+            !msg_res) {
+          return cb(msg_res.error());
+        }
+        return cb(std::move(msg));
+      });
     }
 
     /**
@@ -63,19 +72,24 @@ namespace kagome::network {
      * @param cb to be called, when the message is written, or error happens
      */
     template <typename MsgType>
-    void write(const MsgType &msg, libp2p::basic::Writer::WriteCallbackFunc cb) const {
-      using ProtobufRW = MessageReadWriter<ProtobufMessageAdapter<MsgType>, NoSink>;
+    void write(const MsgType &msg,
+               libp2p::basic::Writer::WriteCallbackFunc cb) const {
+      using ProtobufRW =
+          MessageReadWriter<ProtobufMessageAdapter<MsgType>, NoSink>;
       using UVarRW = MessageReadWriter<UVarMessageAdapter<MsgType>, ProtobufRW>;
 
       // TODO(iceseer) : try to cache this vector
       std::vector<uint8_t> out;
       auto it = UVarRW::write(msg, out);
 
-      gsl::span<uint8_t> data(it.base(), out.size() - std::distance(out.begin(), it));
+      gsl::span<uint8_t> data(it.base(),
+                              out.size() - std::distance(out.begin(), it));
       assert(!data.empty());
 
       read_writer_->write(data,
-                          [self{shared_from_this()}, out{std::move(out)}, cb = std::move(cb)](auto &&write_res) {
+                          [self{shared_from_this()},
+                           out{std::move(out)},
+                           cb = std::move(cb)](auto &&write_res) {
                             if (!write_res) {
                               return cb(write_res.error());
                             }
@@ -86,6 +100,7 @@ namespace kagome::network {
    private:
     std::shared_ptr<libp2p::basic::MessageReadWriter> read_writer_;
   };
+
 }  // namespace kagome::network
 
 #endif  // KAGOME_PROTOBUF_MESSAGE_READ_WRITER_HPP
