+++ test/core/network/types/uvar_adapter_test.cpp
@@ -2,7 +2,7 @@
  * Copyright Soramitsu Co., Ltd. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
-#include "network/helpers/adapters.hpp"
+#include "network/adapters/uvar.hpp"
 
 #include <gmock/gmock.h>
